// src/components/ui/Modals/ModalMainTableData/enums.ts
export enum ENameFieldsInvoiceAddeess {
  company = `company`,
  name = `name`,
  additional = "additional",
  street = "street",
  postalCode = "postalCode",
  address = "address",
  country = "country",
}

export enum ENameFieldsBankData {
  iban = `iban`,
  bic = `bic`,
  bankName = "bankName",
}

export enum ENameFieldsContact {
  fax = `fax`,
  email = `email`,
  additional = "additional",
  birthday = "birthday",
  homepage = "homepage",
}